update dependencies, move config to settings to avoid namespace collision with app.config
