+++ README.md
@@ -26,7 +26,7 @@ This is a strong Flask base/boilerplate focusing on simplicity, while still havi
 #### Libraries
 
 - **Flask:** Flask-SQLAlchemy, Flask-WTF, Flask-Login, Alembic/Flask-Migrate
-- **JS & Frontend:** Bootstrap 4, Gulp, Livereload, and Webpack
+- **JS & Frontend:** Bootstrap 4 (beta 2), Gulp, Livereload, and Webpack
 
 ### Quick Setup
 
@@ -57,7 +57,7 @@ npm run buildprod # minified, prod-ready
 source venv/bin/activate
 python run.py
 
-# Livereload and SASS
+# Livereload and SASS compilation
 gulp
 
 # To recompile bundle.js (if you change app.js), run one the following:
@@ -78,12 +78,12 @@ This boilerplate is ready for MySQL and Postgres, but uses sqlite by default. If
 
 1. Create your database
 2. In `app/config.py` you should comment the sqlite config line (11) and uncomment either mysql (19-20) or postgres (23-24). Make sure that you change the database name and add passwords or anything else you need to the connect string!
-3. Execute either `pip install -r requirements.txt/mysql.txt` or `pip install -r requirements.txt/postgres.txt` (while inside virtual environment) to install the libraries needed for each respective database.
+3. Execute either `pip install -r requirements/mysql.txt` or `pip install -r requirements/postgres.txt` (while inside virtual environment) to install the libraries needed for each respective database.
 
 ### Setting up mail
 *Optional*
 
-A lot of applications tend to use Flask-Mail and suggest to create a gmail address to send email with. While this works fine for small things, it isn't production-ready and Gmail doesn't exist to send transactional email (forgot password, confirm email, etc.). For this reason, my mail setup uses Sparkpost which has 100,000 emails per month for free. I am not sponsored or compensated in any way by Sparkpost. I have never exceeded this monthly limit, and the platform is rich and simple to use. It has reporting and other features that are great for real world applications.
+A lot of applications tend to use Flask-Mail and suggest creating a gmail address to send email with, via smtp. While this works fine for small things, it isn't production-ready and Gmail doesn't exist to send transactional email (forgot password, confirm email, etc.). For this reason, this boilerplate mail setup uses Sparkpost which has 15k emails per month for free (used to be 100k until sometime in 2017). I am not sponsored or compensated in any way by Sparkpost, it's just worked fine for me for sending tranasctional mail free.
 
 1. Create an account on sparkpost.com
 2. In your local environment, set the `SPARKPOST_API_KEY`:
@@ -91,7 +91,7 @@ A lot of applications tend to use Flask-Mail and suggest to create a gmail addre
 3. Update mail config variables in config.py 
 4. All set!
 
-If you use another transactional platform like Mandrill, it should be very simple to drop in here. Feel free to submit a PR with a new branch for any alternative email setups.
+If you use another transactional mail platform, it should be very simple to drop in here by replacing the email send calls with another library or via REST with the requests library. Feel free to submit a PR with a new branch for any alternative email setups.
 
 ### Deploying
 
@@ -100,3 +100,5 @@ In order to deploy, you need to move your entire application & database to a ser
 Basic instructions for a VPS can be found in [a gist here](https://gist.github.com/mcescalante/5db616b9a826605f1df35f79b09cf6f6) but automated scripting & more will be coming soon.
 
 More coming soon!
+
+_Pst: check out the TODO file for a peek at what I would like to make happen soon, but haven't yet_
