+++ package.json
@@ -18,7 +18,7 @@
     "gulp-sass": "^2.3.2"
   },
   "dependencies": {
-    "bootstrap": "^4.0.0-beta.2",
+    "bootstrap": "^4.0.0",
     "jquery": "^3.1.1",
     "popper.js": "^1.12.6",
     "tether": "^1.3.7",
