+++ requirements/base.txt
@@ -1,9 +1,9 @@
-alembic==0.9.6
-certifi==2017.7.27.1
+alembic==0.9.9
+certifi==2018.1.18
 chardet==3.0.4
 click==6.7
 Flask==0.12.2
-Flask-Login==0.4.0
+Flask-Login==0.4.1
 Flask-Migrate==2.1.1
 Flask-Script==2.0.6
 Flask-SQLAlchemy==2.3.2
@@ -11,14 +11,14 @@ Flask-WTF==0.14.2
 gunicorn==19.7.1
 idna==2.6
 itsdangerous==0.24
-Jinja2==2.9.6
+Jinja2==2.10
 Mako==1.0.7
 MarkupSafe==1.0
-python-dateutil==2.6.1
+python-dateutil==2.7.2
 python-editor==1.0.3
 requests==2.18.4
 six==1.11.0
-SQLAlchemy==1.1.14
+SQLAlchemy==1.2.6
 urllib3==1.22
-Werkzeug==0.12.2
+Werkzeug==0.14.1
 WTForms==2.1
